split prompts.py into many smaller files to make it easier to find relevant prompts
