+++ tools/text/prompts/__init__.py
@@ -1,0 +1,5 @@
+from .classify_video import *
+from .lecture_summary import *
+from .parse_image import *
+from .queries import *
+from .search_by_image import *
